config/urls.py
"""
URL configuration.
"""
from django.contrib import admin
from django.http import HttpResponse
from django.urls import path

def health(request):
    return HttpResponse("ok", content_type="text/plain")

urlpatterns = [
    path("admin/", admin.site.urls),
    path("health/", health),
]
