requirements-dev.txt
-r requirements.txt
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
ipython>=8.0.0
black>=24.0.0
flake8>=7.0.0
